feat(judge): [US-004-006] Train the classifier on the train half; held-out runs
scripts/train_classifier.py fine-tunes microsoft/deberta-v3-base (chosen: it
downloaded in 16 s; the local zero-shot backbone stays the fallback) on the
train half only — seed 2026, 10 epochs, lr 2e-5, batch 8, 256 tokens, fp32
(the backbone config declares fp16, which diverges to NaN under AdamW), MPS.
Provenance in docs/runs/finetuned/<dataset>.train.json: hyper-parameters,
per-epoch loss, wall time (emails 215 s, router 1,069 s), hardware, backbone
revision, sha256 of the train rows and of the model's config.json.

scripts/audit_resumable.py --rows <split.json>:<part> judges only those row
indices and records the split, its sha256 and the part in the checkpoint
header; a checkpoint refuses to continue on a different subset.
scripts/arena_run_heldout.sh runs the four datasets that way (email-adversarial
in full). scripts/arena_report.py sets held-out runs aside from the main table
and points to the held-out report; chart labels know the fine-tuned slug.

+++ src/judge_audit/judges/finetuned.py
@@ -65,7 +65,8 @@ def build_predict(model_dir: str, device: str, max_len: int) -> Predict:
     except ImportError as e:
         raise RuntimeError("the finetuned judge needs: pip install 'kunko-judge-audit[nli]'") from e
     tok = AutoTokenizer.from_pretrained(model_dir)
-    model = AutoModelForSequenceClassification.from_pretrained(model_dir).to(device).eval()
+    model = AutoModelForSequenceClassification.from_pretrained(
+        model_dir, dtype=torch.float32).to(device).eval()
 
     def predict(text: str) -> list[float]:
         enc = tok(text, truncation=True, max_length=max_len, return_tensors="pt").to(device)
